refactoring(step3/ui): ui 패키지 view 로 패키지명 변경
