+++ src/main/java/racingcar/view/Result.java
@@ -1,6 +1,4 @@
-package racingcar.ui;
-
-import java.util.List;
+package racingcar.view;
 
 import racingcar.domain.Cars;
